feat: split singe page into routes
Split up the single index.vue file into multiple
routes. The new routes are /about-me, /experience,
 /hire-me, /portfolio. Also changed the default
 layout to have a hamburger menu in on smaller
 screens. The nav is now stuck to the top of the
  screen and is always in view while the footer
  is now at the bottom of the screen.
