+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "vue3snippets.enable-compile-vue-file-on-did-save-code": false
+}
