+++ package.json
@@ -17,6 +17,8 @@
     "nuxt-schema-org": "^2.2.0"
   },
   "dependencies": {
+    "@headlessui/vue": "^1.7.23",
+    "@heroicons/vue": "^2.1.5",
     "@nuxtjs/tailwindcss": "^6.7.0",
     "@tailwindcss/typography": "^0.5.9",
     "@vee-validate/nuxt": "^4.9.6",
