+++ pages/about-me.vue
@@ -1,0 +1,70 @@
+<template>
+  <main>
+    <section
+      id="about-me"
+      class="pb-48 scroll-mt-5 layout-section"
+      aria-labelledby="heading-aboutme"
+    >
+      <div class="max-w-5xl px-5 mx-auto">
+        <h2 id="heading-aboutme" class="title--sm mb-9">About Me</h2>
+        <div class="layout-grid-half">
+          <div>
+            <p class="text--lg">
+              With a proven track record as a Full Stack Developer, Jordan is
+              proficient in JavaScript, HTML, CSS, React, Vue.js, and Node.js.
+              He excels at building intuitive user interfaces that are both
+              aesthetically pleasing and functional.
+            </p>
+            <p class="text--lg">
+              In his most recent role as an OutSystems Developer, Jordan built
+              user-friendly web and mobile apps to streamline complex business
+              processes. He championed low-code development, optimized app
+              UX/UI, and contributed to projects like a mobile banking app and a
+              web application for managing workers' compensation claims.
+            </p>
+            <p class="text--lg">
+              At Upwork, Jordan specialized in developing and customizing
+              Shopify themes and plugins, integrating third-party apps, and
+              ensuring optimal store performance while closely collaborating
+              with clients.
+            </p>
+          </div>
+          <div>
+            <p class="text--lg">
+              During his time at TapOnIt, Jordan gained experience in creating
+              user-friendly web pages and mobile apps, integrating back-end
+              functionality with appealing UI designs. His focus on clean,
+              efficient code resulted in high-quality web designs and
+              applications.
+            </p>
+            <p class="text--lg">
+              Jordan has also completed the Udacity Cloud Developer Nanodegree,
+              enhancing his expertise in AWS services such as Lambda, S3,
+              DynamoDB, and API Gateway, allowing him to design scalable, secure
+              cloud solutions.
+            </p>
+            <p class="text--lg">
+              His diverse skill set spans front-end technologies like React and
+              Vue.js, and back-end development with Node.js and PostgreSQL.
+              Jordan’s commitment to staying updated with the latest
+              technologies ensures innovative solutions for every project.
+            </p>
+            <p class="text--lg">
+              His commitment to staying up-to-date with the latest trends and
+              technologies in the field enables him to deliver innovative
+              solutions and provide valuable insights to drive project success.
+              He constantly seeks new challenges and opportunities to expand his
+              skill set and contributes to impactful projects.
+            </p>
+            <p class="text--lg">
+              Explore his projects or reach out for potential collaborations via
+              the
+              <a href="#portfolio">portfolio</a> or reach out via the
+              <a href="#hire-me">contact form</a>.
+            </p>
+          </div>
+        </div>
+      </div>
+    </section>
+  </main>
+</template>
